Add a reference world with realistic benchmarks
